erreur 403 PUT

+++ controllers/sauce.js
@@ -37,7 +37,6 @@ exports.getOneSauce = (req, res, next) => {
         .catch(error => res.status(404).json({ error: 'Sauce non trouvée' }));
 };
 
-
 // Fonction pour modifier une sauce
 exports.modifySauce = (req, res, next) => {
     // Vérification si une nouvelle image est fournie dans la requête
@@ -45,26 +44,49 @@ exports.modifySauce = (req, res, next) => {
         ...JSON.parse(req.body.sauce),
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     } : { ...req.body };
-    // Mise à jour de la sauce dans la base de données
-    Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
-        .then(() => res.status(200).json({ message: "Sauce modifiée avec succès !" }))
-        .catch(error => res.status(400).json({ error }));
-};
 
-// Fonction pour supprimer une sauce
-exports.deleteSauce = (req, res, next) => {
     // Recherche de la sauce par son identifiant dans la base de données
-    Sauce.findOne({ _id: req.params.id })
+    Sauce.findOne({ _id: req.params.id, userId: req.userData.userId })
         .then(sauce => {
-            // Récupération du nom du fichier image de la sauce
-            const filename = sauce.imageUrl.split('/images/')[1];
-            // Suppression du fichier image de la sauce
-            fs.unlink(`images/${filename}`, () => {
-                // Suppression de la sauce dans la base de données
-                Sauce.deleteOne({ _id: req.params.id })
-                    .then(() => res.status(200).json({ message: "Sauce supprimée avec succès !" }))
-                    .catch(error => res.status(400).json({ error }));
-            });
+            // Vérification si la sauce existe et si l'utilisateur actuel est celui qui a créé la sauce
+            if (!sauce) {
+                return res.status(403).json({ error: 'Vous n\'êtes pas autorisé à modifier cette sauce' });
+            }
+            // Mise à jour de la sauce dans la base de données
+            Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
+                .then(() => res.status(200).json({ message: "Sauce modifiée avec succès !" }))
+                .catch(error => res.status(400).json({ error }));
         })
         .catch(error => res.status(500).json({ error }));
 };
+
+
+
+
+exports.deleteSauce = (req, res, next) => {
+    /* Utilisation de la méthode findOne() du modèle Mongoose qui renvoit la Sauce 
+    ayant le même _id que le paramètre de la requête */
+    Sauce.findOne({ _id: req.params.id })
+      .then((sauce) => {
+        // On verifie que la sauce appartient bien à l'utilisateur
+        if (sauce.userId != req.auth.userId) {
+          res.status(401).json({ message: "Non-autorisé !" });
+        } else {
+          // Séparation du nom du fichier grâce au "/images/"" contenu dans l'url
+          const filename = sauce.imageUrl.split("/images/")[1];
+          // Utilisation de la fonction unlink pour supprimer l'image et suppression de toute la Sauce
+          fs.unlink(`images/${filename}`, () => {
+            Sauce.deleteOne({ _id: req.params.id })
+              .then(() => {
+                res.status(200).json({ message: "Sauce Supprimée !" });
+              })
+              .catch((error) => {
+                res.status(401).json({ error });
+              });
+          });
+        }
+      })
+      .catch((error) => {
+        res.status(500).json({ error });
+      });
+  };
